requirements.txt
streamlit>=1.35.0
pandas>=2.1.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
torch>=2.1.0
torchvision>=0.16.0
